refactor: export constant at utils/index
